change file structure and rename example folder
