+++ sweat_smart/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ExampleConfig(AppConfig):
+class SweatSmartConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'example'
+    name = 'sweat_smart'
